% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_data.R
\name{TestDataGenerator}
\alias{TestDataGenerator}
\title{Test Data Generator}
\description{
This class generates runtimes, and test load and weather data available for the runtimes
}
\section{Usage}{

\preformatted{gen <- TestDataGenerator$new()
result <- gen$next_runtime(lag=20)
runtime <- result$runtime
data <- result$data
}
}

